Move file watcher into its own file
