+++ src/main.rs
@@ -10,6 +10,7 @@ pub mod renderer;
 pub mod table;
 pub mod text;
 pub mod utils;
+mod watcher;
 
 use crate::image::Image;
 use crate::interpreter::HtmlInterpreter;
@@ -41,10 +42,6 @@ use anyhow::Context;
 use copypasta::{nop_clipboard::NopClipboardContext as ClipboardContext, ClipboardProvider};
 #[cfg(feature = "x11")]
 use copypasta::{ClipboardContext, ClipboardProvider};
-use notify::{
-    event::{EventKind, ModifyKind},
-    RecommendedWatcher, RecursiveMode, Watcher,
-};
 
 use winit::event::ModifiersState;
 use winit::event::{ElementState, MouseButton};
@@ -68,7 +65,6 @@ use std::sync::mpsc;
 use std::sync::mpsc::channel;
 use std::sync::Arc;
 use std::sync::Mutex;
-use std::time::Duration;
 
 pub enum InlyneEvent {
     LoadedImage(String, Arc<Mutex<Option<ImageData>>>),
@@ -185,69 +181,6 @@ fn root_filepath_to_vcs_dir(path: &Path) -> Option<PathBuf> {
 }
 
 impl Inlyne {
-    // TODO: doesn't follow watching file after file changes. Need to coordinate that
-    pub fn spawn_watcher(&self) {
-        // Create a channel to receive the events.
-        let (watch_tx, watch_rx) = channel();
-
-        // Create a watcher object, delivering raw events.
-        // The notification back-end is selected based on the platform.
-        let mut watcher = RecommendedWatcher::new(watch_tx, notify::Config::default()).unwrap();
-
-        // Add the file path to be watched.
-        let event_proxy = self.event_loop.create_proxy();
-        let file_path = self.opts.file_path.clone();
-        std::thread::spawn(move || {
-            watcher
-                .watch(&file_path, RecursiveMode::NonRecursive)
-                .unwrap();
-
-            loop {
-                let event = match watch_rx.recv() {
-                    Ok(Ok(event)) => event,
-                    Ok(Err(err)) => {
-                        log::warn!("File watcher error: {}", err);
-                        continue;
-                    }
-                    Err(err) => {
-                        log::warn!("File watcher channel dropped unexpectedly: {}", err);
-                        break;
-                    }
-                };
-
-                log::debug!("File event: {:#?}", event);
-                match event.kind {
-                    EventKind::Remove(_) | EventKind::Modify(ModifyKind::Name(_)) => {
-                        // Some editors may remove/rename the file as a part of saving.
-                        // Reregister file watching in this case
-                        log::debug!("File may have been renamed/removed. Falling back to polling");
-
-                        let mut delay = Duration::from_millis(10);
-                        loop {
-                            std::thread::sleep(delay);
-                            delay = Duration::from_millis(100);
-
-                            let _ = watcher.unwatch(&file_path);
-                            if watcher
-                                .watch(&file_path, RecursiveMode::NonRecursive)
-                                .is_ok()
-                            {
-                                log::debug!("Sucessfully re-registered file watcher");
-                                let _ = event_proxy.send_event(InlyneEvent::FileReload);
-                                break;
-                            }
-                        }
-                    }
-                    EventKind::Modify(_) => {
-                        log::debug!("Reloading file");
-                        let _ = event_proxy.send_event(InlyneEvent::FileReload);
-                    }
-                    _ => {}
-                }
-            }
-        });
-    }
-
     pub fn new(opts: Opts) -> anyhow::Result<Self> {
         let keycombos = KeyCombos::new(opts.keybindings.clone())?;
 
@@ -823,7 +756,7 @@ fn main() -> anyhow::Result<()> {
     let opts = Opts::parse_and_load_from(args, config)?;
     let inlyne = Inlyne::new(opts)?;
 
-    inlyne.spawn_watcher();
+    watcher::spawn_watcher(&inlyne);
     inlyne.run();
 
     Ok(())
